refactor(scheduler): merge solvers/v2 into solvers, delete dead code
The v2 package contents move up to solvers/. Package declaration
changes from v2 to solvers; v2.X references in job_solver.go and the
three action callers become bare names or solvers.X.

Also deletes the legacy PodAccumulatedScenarioBuilder and its test —
both unused since Phase 2 inlined the accumulation logic into the v2
generator.

The v2 directory is gone. solvers/ is the production solver package
with the three-interface stack (Generator, Simulator, Validator) and
the JobSolver wrapper that adapts it to the action call shape.
