+++ pkg/scheduler/actions/reclaim/reclaim.go
@@ -24,7 +24,6 @@ import (
 
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/actions/common"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/actions/common/solvers"
-	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/actions/common/solvers/v2"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/actions/utils"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/common_info"
 	"github.com/kai-scheduler/KAI-scheduler/pkg/scheduler/api/podgroup_info"
@@ -111,7 +110,7 @@ func (ra *reclaimAction) attemptToReclaimForSpecificJob(
 	ssn.OnJobSolutionStart()
 
 	feasibleNodes := common.FeasibleNodesForJob(maps.Values(ssn.ClusterInfo.Nodes), reclaimer)
-	validator := v2.LegacyValidator(ssn, "reclaim", ssn.ReclaimScenarioValidatorFn)
+	validator := solvers.LegacyValidator(ssn, "reclaim", ssn.ReclaimScenarioValidatorFn)
 	solver := solvers.NewJobsSolver(
 		feasibleNodes,
 		validator,
